TEST Product CRUD with Authentication

+++ ProductWebAPI/Controllers/ProductImageController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using ProductWebAPI.Models;
 using ProductWebAPI.Services.DocumentService;
 using ProductWebAPI.Services.DomainService;
+using static ProductWebAPI.Utilities;
 
 namespace ProductWebAPI.Controllers
 {
@@ -11,13 +12,21 @@ namespace ProductWebAPI.Controllers
     [Authorize]
     public class ProductImageController : ControllerBase
     {
+        #region global fields
         private readonly IDocumentManager _documentManager;
         private readonly IProductImageService _productImageService;
+        #endregion
+
+        #region CTOR
         public ProductImageController(IDocumentManager documentManager, IProductImageService productImageService)
         {
             _documentManager = documentManager;
             _productImageService = productImageService;
         }
+
+        #endregion
+
+        #region Action Results
         [HttpPost]
         public async Task<ActionResult> Create([FromForm] NewProductImageModel model)
         {
@@ -39,13 +48,30 @@ namespace ProductWebAPI.Controllers
         public async Task<ActionResult> Delete(int productImageId)
         {
             SaveResult saveResult = new SaveResult();
-            var model = await _productImageService.GetById(productImageId);
+            string? userId = GetUserId();
+
+            if (userId == null) return Unauthorized();
+
+            var model = await _productImageService.GetById(productImageId, userId);
 
             if(!String.IsNullOrEmpty(model.FileNameGuid) && await _documentManager.Delete(model.FileNameGuid)) { 
                 saveResult =await _productImageService.Delete(productImageId);
             }
            return saveResult.IsSuccess ? Ok() : BadRequest(saveResult.Message);
         }
+        #endregion
+
+        #region private methods
+        private string? GetUserId()
+        {
+            if (User.Claims == null)
+            {
+                return null;
+            }
+
+            return User.Claims.ExtractUserId(ClaimTypeEnum.sid.ToString());
+        }
+        #endregion
     }
 }
